realtime-rec-engine-v2/requirements.txt
# Core ML/AI Libraries
torch==2.1.0
torch-audio==2.1.0
torchvision==0.16.0
scann==1.9.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0

# Data Processing & Streaming
apache-kafka==2.3.0
confluent-kafka==2.2.0
avro==1.11.3
pyarrow==13.0.0
parquet==1.3.1

# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
httpx==0.25.1

# Database & Cache
redis==5.0.1
redis-py-cluster==2.1.3
postgresql==0.0.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1

# Feature Store & MLOps
feast==0.38.0
mlflow==2.7.1
bentoML==1.1.0
whylogs==1.3.13

# Monitoring & Observability
prometheus-client==0.18.0
grafana-api==1.0.3
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
structlog==23.2.0
sentry-sdk==1.32.0

# Infrastructure & DevOps
kubernetes==28.1.0
helm==0.0.1
docker==6.1.3
boto3==1.29.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0

# Testing & Quality
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.17.0
black==23.10.1
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0

# Utilities
click==8.1.7
pyyaml==6.0.1
python-dotenv==1.0.0
jinja2==3.1.2
requests==2.31.0
aiohttp==3.9.0
asyncio-mqtt==0.13.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Performance & Optimization
orjson==3.9.10
lz4==4.3.2
xxhash==3.4.1
numba==0.58.1

# Distributed Computing
ray==2.8.0
dask==2023.11.0
celery==5.3.4

# Configuration Management
hydra-core==1.3.2
omegaconf==2.3.0
